\docType{package}
\name{ymq-package}
\alias{ymq-package}
\alias{ymq}

\title{Year Months and Quarters}

\description{
Operators to manipulate integers representing year months in the form YYYYMM, and year quarters in the form YYYYQQ.
}

\seealso{
\url{https://github.com/marberts/yqm}
}

\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com}
}